Move DEFAULT_WMS_VERSION to a common source module
