+++ src/ol/index.js
@@ -6,12 +6,6 @@ import webgl from './webgl.js';
 
 
 /**
- * @type {string} Default WMS version.
- */
-export var DEFAULT_WMS_VERSION = '1.3.0';
-
-
-/**
  * @type {boolean} Include debuggable shader sources.  Default is `true`.
  *     This should be set to `false` for production builds.
  */
@@ -128,7 +122,6 @@ export default {
   nullFunction: nullFunction,
   inherits: inherits,
   VERSION: VERSION,
-  DEFAULT_WMS_VERSION: DEFAULT_WMS_VERSION,
   DEBUG_WEBGL: DEBUG_WEBGL,
   HAS_WEBGL: HAS_WEBGL,
   WEBGL_MAX_TEXTURE_SIZE: WEBGL_MAX_TEXTURE_SIZE,
