+++ src/ol/source/common.js
@@ -1,0 +1,4 @@
+/**
+ * @type {string} Default WMS version.
+ */
+export var DEFAULT_WMS_VERSION = '1.3.0';
